+++ tools/compiler/build.sh
@@ -1,1 +1,2 @@
+go get
 GOOS=js GOARCH=wasm go build -o static/main.wasm compiler
